requirements.txt
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
fastapi==0.113.0
h11==0.14.0
idna==3.8
iso8601==1.1.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.9.0
pydantic_core==2.23.2
pypika-tortoise==0.1.6
pytz==2024.1
sniffio==1.3.1
starlette==0.38.4
tortoise-orm==0.21.6
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
